% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateAbundances.R
\name{aggregateAbundances}
\alias{aggregateAbundances}
\title{Aggregate TreeCo Species Abundance Table per Group}
\usage{
aggregateAbundances(
  tree.data = NULL,
  site.data = NULL,
  spp.group = "species.correct",
  site.group = "ordem",
  cols.to.sum = c("N", "AB", "DA", "DoA", "DR", "DoR", "F", "IVC", "IVI"),
  cols.to.paste = c("ordem", "refID", "SiteCode", "SubSite", "Name_submitted", "family",
    "genus", "epiteth", "infra.epiteth", "authorship", "species.correct", "taxon.rank",
    "family.original", "genus.original", "species.original", "author.original", "notes",
    "accession", "coletor", "number", "DetBy", "DetDate", "obs"),
  effort = "effort_ha"
)
}
\arguments{
\item{tree.data}{a data frame with tree abundance data.}

\item{site.data}{a data frame with site and plot metadata (optional).}

\item{spp.group}{a column with the species names to be used for aggregation.}

\item{site.group}{a column with the site or plot codes/names to be used for aggregation.}

\item{cols.to.sum}{a vector with the names of the columns in `tree.data` to
be summed over species and groups (generally the numerical variables)}

\item{cols.to.paste}{a vector with the names of the columns in `tree.data` to
be be pasted over species and groups (generally the categorical variables)}

\item{effort}{the name of the columns in `site.data` containg the effort of
each inventory in hectares}
}
\value{
a data frame with the aggregated information per species name per
  group (e.g. plot ID or site code)
}
\description{
by ordem or SiteCode
}
\details{
Users ...
}
\author{
Renato A. F. de Lima
}
