% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract_money.R
\name{dollars}
\alias{dollars}
\title{Creates dollars column that multiplies conversion value}
\usage{
dollars(x, pattern, conversion)
}
\arguments{
\item{x}{Tree data}

\item{pattern}{Pattern to match benefit type}

\item{conversion}{Conversion value}
}
\description{
Creates dollars column that multiplies conversion value
}
