% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.cat2mean_EG}
\alias{.cat2mean_EG}
\title{Convert Ecological Groups to Ordinal Values}
\usage{
.cat2mean_EG(x)
}
\arguments{
\item{x}{a vector of ecologial groups}
}
\value{
the character `x` with the values for the categories
}
\description{
Aiming to obtain ordinal variables from ecological
  groups (e.g. pioneer, early secondary, etc.) categories are
  converted to numbers as follows:
  - pioneer: 1
  - light_demanding or shade_intolerant: 1.5
  - early_secondary or heliofita_to_mesofita: 2
  - mesofita: 2.5
  - late_secondary or shade_tolerant: 3
  - ciofita or climax: 4
  
 Variations of these general categories are dealed internally to
 maximize the convertion of categories into numbers. For instance,
 categories 'heliofita_to_mesofita' are treated as being
 'early_secondary'.
}
\keyword{internal}
