% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco_data}
\alias{eco_data}
\title{Eco data}
\format{A data frame with 34164 rows and 5 variables:
\describe{
  \item{key}{species code}
  \item{region}{region code}
  \item{dbh_range}{a range of dbh values in cm}
  \item{benefit_value}{tree benefit values}
  \item{benefit}{benefit type}
  ...
}}
\source{
\url{https://www.itreetools.org}
}
\usage{
eco_data
}
\description{
A dataset containing the benefits and other attributes.
}
\keyword{datasets}
