% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_guess.R
\name{eco_guess}
\alias{eco_guess}
\title{Guess common or botanical names}
\usage{
eco_guess(x, guess)
}
\arguments{
\item{x}{A vector containing "common" or "botanical" names.}

\item{guess}{The missing field, either "common" or "botanical".}
}
\description{
This function tries to guess the botanical or common names of
trees for users with only one of these fields, not both. Since
\code{eco_run_all} attempts to find matches in both common and botanical,
running this is required for users that are missing one of these fields.
}
\examples{
eco_guess("common fig", "botanical")

eco_guess("pinus eldarica", "common")

}
