% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_run.R
\name{eco_run}
\alias{eco_run}
\title{Run eco benefits for a single tree}
\usage{
eco_run(common, dbh, region)
}
\arguments{
\item{common}{the common name of the tree}

\item{dbh}{the dbh value in inches}

\item{region}{region code, see \code{money} dataset for region codes}
}
\description{
Run eco benefits for a single tree
}
