% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_run_all.R
\name{eco_run_all}
\alias{eco_run_all}
\title{Run eco benefits for an entire tree inventory}
\usage{
eco_run_all(data, species_col, dbh_col, region)
}
\arguments{
\item{data}{path to csv file containing tree inventory}

\item{species_col}{the name of the column containing common names of species}

\item{dbh_col}{the name of the column containing dbh values}

\item{region}{region code, see \code{species} or \code{benefits}}
}
\description{
Run eco benefits for an entire tree inventory
}
