% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract_money.R
\name{extract_money}
\alias{extract_money}
\title{Extracts money benefits}
\usage{
extract_money(tree_data, money_data)
}
\arguments{
\item{tree_data}{Tree dataset}

\item{money_data}{Money dataset}
}
\description{
Extracts money benefits
}
