% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract_money.R
\name{find_conversion}
\alias{find_conversion}
\title{Finds the conversion value to transform benefits to dollars}
\usage{
find_conversion(x, pattern)
}
\arguments{
\item{x}{Money data}

\item{pattern}{Pattern to match benefit type}
}
\description{
Finds the conversion value to transform benefits to dollars
}
