% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functcomp.treeco.R
\name{functcomp.treeco}
\alias{functcomp.treeco}
\title{Calculate Trait Community Weighted Means}
\usage{
functcomp.treeco(x, a, CWM.type = "all", bin.num = NULL, CWM.sd = TRUE)
}
\arguments{
\item{x}{matrix or data frame containing the functional traits.}

\item{a}{matrix or data frame containing the abundances of the taxa listed in x.}

\item{CWM.type}{character string indicating how nominal, binary and ordinal
traits should be handled. Default to "all".}

\item{bin.num}{vector indicating binary traits to be treated as continuous.}

\item{CWM.sd}{should the weighted standard deviation of the numerical traits de
returned? Default to TRUE.}
}
\description{
This function is adapted from the function `functcomp()` from
  package `FD` (Laliberté et al. 2014). It returns for each trait the
  community-level weighted means (CWM), which can be used to describe the
  functional composition of communities.
}
\details{
This function uses most of the codes from `FD::functcomp()`, so the
  help and details of this function should be consulted.
  
  The main differences of `functcomp.treeco()` from `FD::functcomp()` are:
   -- FINALIZAR --
}
\references{
Laliberté, E., Legendre, P., and B. Shipley. (2014). FD: measuring
 functional diversity from multiple traits, and other tools for functional
 ecology. R package version 1.0-12.
}
\seealso{
\link[FD]{functcomp}
}
