% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCWM.R
\name{getCWM}
\alias{getCWM}
\title{Obtain Community Weighted Means}
\usage{
getCWM(
  tree.data = NULL,
  trait.data = NULL,
  group = "ordem",
  spp.name = "Name_submitted",
  tax.rank = "taxon.rank",
  trait.list = c("wsg_gcm3", "MaxHeight_m", "SeedMass_g", "extinction", "endemism",
    "LeafArea", "LeafType", "dispersal.syndrome", "succesional.group",
    "ecological.group"),
  indets = c("unidentified"),
  rm.flora = TRUE,
  treeco.cwm = TRUE,
  reduce.cats = TRUE,
  habit = "habito",
  life.form = "life.form",
  ab.metric = "counts"
)
}
\arguments{
\item{tree.data}{the data frame with tree abundance data.}

\item{trait.data}{the data frame with tree species trait data.}

\item{group}{vector with the the column that should be used to aggregate the
data. Default to 'ordem'.}

\item{spp.name}{the name of the columns containing the (morpho)species names.
Default to the TreeCo column 'species.correct'.}

\item{tax.rank}{the name of the columns containing the name taxonomic resolution.
Default to the TreeCo column 'taxon.rank'.}

\item{trait.list}{a vector with the names of the traits should be included in
the calculations.}

\item{indets}{a vector of characters containing the unidentified classes in
`tax.rank`. Default to the TreeCo classes 'unidentified', 'family' and
'genus'.}

\item{rm.flora}{should the records of (morpho)species not found within the
main sampling but found in the floristics or natural regeneration of the
same survey. Default to TRUE.}

\item{treeco.cwm}{should the palms, ferns and/or shrubs be removed prior to
the calculation of the CWM of wood density, potential height and seed mass?
Default to TRUE.}

\item{reduce.cats}{should the categories of dispersal and ecological groups
be simplified previous to the analysis? Default to TRUE.}

\item{habit}{the name of the columns containing the species habit information
(i.e. 'tree', 'shrub', 'others')}

\item{life.form}{the name of the columns containing the species life form information
(i.e. 'woody_tree', 'woody_vines_and_subshrubs', 'palm', 'tree_fern',
'succulent_tree' or 'palmoids').}

\item{ab.metric}{the abundance metric that should be used to weight the
community-level trait means: 'counts' or 'biomass'. Default to 'counts'.}
}
\value{
a data frame with the CWMs for each trait X group combination.
}
\description{
Calculate the Community Weighted Means (CWM), particularly for
  data in the TreeCo format.
}
\details{
The function .... TO BE FINISHED

If the argument `treeco.cwm` is TRUE: 
- Information on species 'habit' and 'life.form' are necessary for filtering
somes species out of the CWM from certain traits.
- The CWM for each trait is calculated with a different set of species. For
Wood Density, CWM excluded non woody species (e.g. palms and ferns). For Max.
Height, CWM excludes shrubs (H and/or DBH typically <4m and 10 cm). Seed mass
was calculated without tree ferns."

If the argument `treeco.cwm` is TRUE: 
- The function reduces the dispersal syndrome to abiotic (anemochoric and
autochoric) and zoochoric; hydrochory is disregarded
}
\author{
Renato A. F. de Lima
}
