% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indetSpecies.R
\name{indetSpecies}
\alias{indetSpecies}
\title{Number of Unidentifications per Survey}
\usage{
indetSpecies(
  tree.data = NULL,
  group = "ordem",
  spp.name = "species.correct",
  tax.rank = "taxon.rank",
  indets = c("unidentified", "family", "genus"),
  not.indets = c("species", "variety", "subspecies"),
  rm.flora = TRUE
)
}
\arguments{
\item{tree.data}{the data frame with tree abundance data in the TreeCo format}

\item{group}{vector with the the column that should be used to aggregate the
data. Default to 'ordem'.}

\item{spp.name}{the name of the columns containing the (morpho)species names.
Default to the TreeCo column 'species.correct'.}

\item{tax.rank}{the name of the columns containing the name taxonomic resolution.
Default to the TreeCo column 'taxon.rank'.}

\item{indets}{a vector of characters containing the unidentified classes in
`tax.rank`. Default to the TreeCo classes 'unidentified', 'family' and
'genus'.}

\item{not.indets}{a vector of characters containing the identified classes in
`tax.rank`. Default to the TreeCo classes 'species', 'variety' and
'subspecies'.}

\item{rm.flora}{should the records of (morpho)species not found within the
main sampling but found in the floristics or natural regeneration of the
same survey be included? Default to TRUE.}
}
\value{
a data frame with number of individuals (N) and of taxa (S) for the each group and
the percentage of N and S unidentified at different taxonomic resolutions.
}
\description{
Calculate the number of individuals and species that are
not identified at different taxonomic levels (e.g. family, genus).
}
\details{
The function returns the percentage of individuals and species that
  remained unidentified until a certain taxonomic rank, which is provided by
  the user. *The order of the vector provided in `indets` must be from the lowest
  to the highest resolution within the unidentification levels*.
  
  The taxonomic ranks include in TreeCo are, in descreasing order
  of taxonomic resolution: 'variety'/'subspecies', 'species', 'genus',
  'family', 'unidentified', 'dead' and 'correction'. The class 'unidentified'
  include living trees for which not even a family name was assigned. The
  class 'correction' is restricted to the Minas Gerais state inventory data.
  This class enter the computations of the total number of individuals in the
  sample but not in the totoal number of taxa.
  
  The proportions are cumulative and they respect the order of taxonomic
  levels provided by the user. By default, this means that proportions at
  genus level include the identifications at the genus, family and
  unidentified classes, proportions at family level include the
  identifications at family and unidentified resolutions. 
  
  Some TreeCo surveys contain extra species lists from outside the sampling
  plots/transects. These lists are often obtaining from floristic surveys in
  the same sites where the sampling was conducted. Other surveys stored in
  TreeCo present the records for species present in the adult and natural
  regeneration components of the forest. If a species is present in the
  floristic survey but not in the plots, the species enters the database with
  an abundance value equal to zero. Similarly, if a species is present in the
  natural regeneration but not in the adult component, the species enters the
  database with an abundance between zero and one. This records are excluded by
  default, suing the argument `rm.flora`.
}
\seealso{
\link[TreeCo]{traitCover}
}
\author{
Renato A. F. de Lima
}
