% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{money}
\alias{money}
\title{Money dataset}
\format{A data frame with 16 rows and 11 variables:
\describe{
  \item{region_code}{Region code.}
  \item{region_name}{Region name associated with region code. This is meant to be human readable for users to pick the appropriate region code.}
  \item{electricity_kwh_to_currency}{Electricity in kwh to dollar amount.}
  \item{natural_gas_kbtu_to_currency}{Natural gas in kbtu to dollar amount.}
  \item{h20_gal_to_currency}{H20 in gallons to dollar amount.}
  \item{co2_lb_to_currency}{CO2 in lbs to dollar amount.}
  \item{o3_lb_to_currency}{O3 in lbs to dollar amount.}
  \item{nox_lb_to_currency}{NOX in lbs to dollar amount.}
  \item{pm10_lb_to_currency}{PM10 in lbs to dollar amount.}
  \item{sox_lb_to_currency}{SOX in lbs to dollar amount.}
  \item{voc_lb_to_currency}{VOC in lbs to dollar amount.}
  ...
}}
\source{
\url{https://www.itreetools.org}
}
\usage{
money
}
\description{
A dataset containing benefit to currency conversion values. This dataset is how benefit values get converted to dollar amounts.
}
\keyword{datasets}
