% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAbundances.R
\name{prepAbundances}
\alias{prepAbundances}
\title{Prepare TreeCo Species Abundance Table}
\usage{
prepAbundances(
  tree.data = NULL,
  spp.name = "species.correct",
  site.name = "SiteCode",
  plot.id = "ordem",
  record.id = "RecordID",
  tax.rank = "taxon.rank",
  N.name = "N",
  AB.name = "AB",
  obs.name = "obs",
  linhas = NULL,
  rm.sites = NULL,
  rm.plots = NULL,
  rm.tax = c("liana", "bamboo", "herb", "exclude"),
  rm.dead = TRUE,
  rm.cf = TRUE,
  rm.infra = TRUE,
  replace.synonym = TRUE,
  rm.spp = NULL,
  rm.extra = TRUE,
  minN = 1
)
}
\arguments{
\item{tree.data}{the data frame with tree abundance data.}

\item{spp.name}{= "species.correct"}

\item{site.name}{= "SiteCode"}

\item{plot.id}{= "ordem"}

\item{record.id}{= "RecordID"}

\item{tax.rank}{= "taxon.rank"}

\item{N.name}{= "N"}

\item{AB.name}{= "AB"}

\item{obs.name}{= "obs"}

\item{linhas}{= NULL}

\item{rm.sites}{= NULL}

\item{rm.plots}{= NULL}

\item{rm.tax}{= c("liana","bamboo","herb","exclude")}

\item{rm.dead}{logical. Should dead individuals be removed? Defaults to TRUE}

\item{rm.cf}{= FALSE}

\item{rm.infra}{= FALSE}

\item{replace.synonym}{= FALSE}

\item{rm.spp}{character. A list of species names to be removed from the
abundance data.}

\item{rm.extra}{logical. Should the additional abundance data for
size cutoffs different than the TreeCo preffered cutoff (e.g. DAP>3.2cm,
DAS>3.2 inteade of DAP>5cm) etc. be removed? Defaults to TRUE.}

\item{minN}{= 1}
}
\value{
a named list containing new data frame and the notes of the
  operations selected and applied to obtain the new data frame.
}
\description{
To be included
}
\details{
User needs to choose the operations he wants to execute
  over the original species abundance dataset
}
\author{
Renato A. F. de Lima
}
