% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTraits.R
\name{prepTraits}
\alias{prepTraits}
\title{Prepare TreeCo Species Trait Table}
\usage{
prepTraits(
  trait.data = NULL,
  spp.name = "Name_submitted",
  tax.rank = "taxon.rank",
  trait.status = "status",
  trait.list = c(WSG = "wsg_gcm3", H = "MaxHeight_m", DBH = "MaxDbh_cm", MinLL =
    "MinLeafLength_cm", MinLW = "MinLeafWidth_cm", MaxLL = "MaxLeafLength_cm", MaxLW =
    "MaxLeafWidth_cm", LA = "LeafArea", SM = "SeedMass_g", EXT = "extinction", END =
    "endemism", LA = "LeafArea", LT = "LeafType", DS = "dispersal.syndrome", EG =
    "ecological.group"),
  rm.dead = TRUE,
  rm.indet = TRUE,
  rm.dup = TRUE
)
}
\arguments{
\item{trait.data}{the data frame with the raw TreeCo trait table.}

\item{spp.name}{character. The column name containing the species names.
Defaults to "Name_submitted"}

\item{tax.rank}{= "taxon.rank"}

\item{trait.status}{character. The column name containing the entry status.
Defaults to "status"}

\item{trait.list}{a named vector with the column names of the traits should
be included in the calculations.}

\item{rm.dead}{logical. Should dead individuals be removed? Defaults to TRUE}

\item{rm.indet}{logical. Should unidentified individuals be removed? Defaults
to TRUE}

\item{rm.dup}{logical. Should the duplicated or possibly duplicated entries
(of wood density) be removed? Defaults to TRUE}
}
\value{
a named list containing the edited and averaged trait data at the
  different taxonomic levels and the notes of the operations applied
  to obtain the new data frame.
}
\description{
Clean trait records, get derived information and average the
species trait information at family, genus, species and infraspecific levels.
}
\details{
To be finished

For obtaining the 'Potential Adult Height' (MaxHeight_m) and
'Maximum stem diameter' (MaxDbh_cm), if an interval is provided for
traits 'H' and 'DBH' (e.g. '10_12') only the higher value is used
for calculations (12 in this case). The 90% quantile of maximum 'H'
and 'DBH' distributions are used to obtain 'MaxHeight_m' and
'MaxDbh_cm', respectivelly. For all other numerical values, if an
interval is provided (e.g. '0.6_0.7') only the mean of the interval
is used (i.e. 0.65) and the mean of the trait distribution is used.

#Explain the means for Leaf, Fruit and Seed Length and Width for simple and compound, with and without intervals or extreme values

For petiole or petiolule length, the description "sessile" is
considered to be 0.01 cm and the descriptions 'subsessile','short
petiolilate','sessile or short petiolate','sessile or subsessile'
are considered as 0.1 cm.

Missing leaf type ('simple' or 'compound') are inputated for some
genera know to have constant leaf types (e.g. Eugenia, Casearia or
Miconia). Unifoliolate compound leafs are atread as simple leaves.
}
\author{
Renato A. F. de Lima & Lucie Zinger
}
