% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refAuthors.R
\name{refAuthors}
\alias{refAuthors}
\title{Get Author Names}
\usage{
refAuthors(
  ref.ids = NULL,
  ref.data = NULL,
  ids.coluna = "refID",
  auth.coluna = "author",
  adv.coluna = "advisor",
  fuzzy = FALSE,
  max.dist = 0.025
)
}
\arguments{
\item{ref.ids}{a vector with the list of TreeCo refIDs}

\item{ref.data}{the data frame with TreeCo reference database}

\item{ids.coluna}{the name of the column in 'ref.data' containing the
TreeCo refIDs. Defaults to 'refID'.}

\item{auth.coluna}{the name of the column in 'ref.data' containing the
author names. Defaults to 'author'.}

\item{adv.coluna}{the name of the column in 'ref.data' containing
the advisor name in the case of thesis and dissertations.
Defaults to 'advisor'.}

\item{fuzzy}{logical. Should authors with different but close
spelling be put together (may take time for large lists of
authors). Defaults to FALSE.}

\item{max.dist}{numerical. The maximum distance between strings to
put authors names together.}
}
\description{
Function to extract and tabulate the names of authors 
based on a list of reference IDs and the TreeCo standard reference 
table
}
\details{
If the argument 'ref.IDs' is empty, all references in
  'ref.data' will be used to get the count of author names
  
  Distances between authors names are performed using the
  `stringdistmatrix()` function from package stringdist and using
  the method 'jw' (i.e. Jaro-Winkler) for for distance calculation.
  For large selections of references, this may take a while.
}
\author{
Renato A. F. de Lima
}
