% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refInfo.R
\name{refInfo}
\alias{refInfo}
\title{Extract Information from References}
\usage{
refInfo(
  ref.data = NULL,
  ref.name = "Reference",
  ref.id = "refID",
  ref.type = NULL
)
}
\arguments{
\item{ref.data}{the data frame with references in the TreeCo format}

\item{ref.name}{the name of the columns containing the reference. Default to 
the TreeCo column 'Reference'.}

\item{ref.id}{the name of the columns containing the reference. Default to 
the TreeCo column 'refID'.}

\item{ref.type}{a vector containing the type of reference (i.e. journal
article, book, book chapter, monography). If not provide, it assumes that
all (missing) references are journal articles.}
}
\value{
a data frame with number with the reference information in each
  column.
}
\description{
Function the compile different information (e.g. authors, year 
of publication) from a reference in the standard TreeCo format
}
\details{
The standard TreeCo format for the references is the following:
Author(s). Year. Title. Publication name volume (issue): start page-endpage.
The separation between multiple authors must be done using semi-colons,
ampersand (i.e. '&'), a vertical pipe (i.e. '|')or the word 'and' surrounded
by spaces in English, Portuguese, Spanish, German or Latin (i.e. ' and ', 
' e ' ', ' y ', ' und ', ' et ').
}
\seealso{
\link[TreeCo]{}
}
\author{
Renato A. F. de Lima
}
