% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Species dataset}
\format{A data frame with 3178 rows and 12 variables:
\describe{
  \item{id}{An id variable.}
  \item{species_code}{Species codes.}
  \item{scientific_name}{Scientific tree names.}
  \item{common_name}{Common tree names.}
  \item{tree_type}{Tree types.}
  \item{spp_value_assignment}{Another tree code for linking trees to benefits.}
  \item{species_rating_percent}{to be determined...}
  \item{basic_price_sq_in}{to be determined...}
  \item{palm_trunk_cost_ft}{to be determined...}
  \item{replacement_cost}{to be determined...}
  \item{t_ar_sq_inches}{to be determined...}
  \item{species_region}{Tree region.}
  ...
}}
\source{
\url{https://www.itreetools.org}
}
\usage{
species
}
\description{
A dataset containing species data and other attributes.
}
\keyword{datasets}
