% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{Species data}
\format{A data frame with 3178 rows and 11 variables:
\describe{
  \item{species_code}{species code}
  ...
}}
\source{
\url{https://www.itreetools.org}
}
\usage{
species_data
}
\description{
A dataset containing species data and other attributes
}
\keyword{datasets}
