% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesList.R
\name{speciesList}
\alias{speciesList}
\title{Prepare List of Species Names for Query}
\usage{
speciesList(
  x = NULL,
  tax.name = "suggestedName",
  spp.name = "species.correct",
  fam.name = "family",
  genus.name = "genus",
  tax.rank = "taxon.rank",
  tax.status = "scientificNameStatus",
  out.names = c("Name_submitted", "family", "genus", "species.correct", "taxon.rank")
)
}
\arguments{
\item{x}{a data frame with, at least, the cleaned and validated taxa names}

\item{tax.name}{name of the column containing the cleaned and validate taxa
names. Default to 'suggestedName'}

\item{spp.name}{name of the column containing the original species name, used
to obtain the cleaned and validated names. Default to 'species.correct'.}

\item{fam.name}{name of the column containing the taxon family. Default to
'family'.}

\item{genus.name}{name of the column containing the taxon genus. Default to
'genus'.}

\item{tax.rank}{name of the column containing the taxon rank. Default to
'taxon.rank'.}

\item{tax.status}{name of the column containing the taxon status. Default to
'scientificNameStatus'.}

\item{out.names}{vector of column names in the output table for 'tax.name',
'fam.name', 'genus.name', 'spp.name' and 'tax.rank', in this specific
order.}
}
\value{
a data frame with ...
}
\description{
From a list of names already cleaned and validated, the function
prepare all fields necessary to obtain species trait data from the TreeCo
database.
}
\details{
The function ...
}
\seealso{
\link[TreeCo]{traitCover}
}
\author{
Renato A. F. de Lima
}
