% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmoney}
\alias{tmoney}
\title{tmoney dataset}
\format{A data frame with 144 rows and 4 variables:
\describe{
  \item{region}{Region code}
  \item{region_name}{Region name associated with region code}
  \item{conversion}{Conversion description}
  \item{value}{Value in dollars}
  ...
}}
\source{
\url{https://www.itreetools.org}
}
\usage{
tmoney
}
\description{
A dataset containing benefit to currency conversion values. This dataset is how benefit values get converted to dollar amounts.
}
\keyword{datasets}
