% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitCover.R
\name{traitCover}
\alias{traitCover}
\title{Assess Trait Species Coverage}
\usage{
traitCover(
  tree.data = NULL,
  trait.data = NULL,
  group = "ordem",
  spp.name = "Name_submitted",
  trait.list = c("wsg_gcm3", "MaxHeight_m", "SeedMass_g", "extinction", "endemism",
    "LeafArea", "LeafType", "dispersal.syndrome", "ecological.group"),
  rm.flora = TRUE,
  trait.miss = 3
)
}
\arguments{
\item{tree.data}{the data frame with tree abundance data in the TreeCo format}

\item{trait.data}{the data frame with edited and averaged trait data in the
TreeCo format}

\item{group}{vector with the the column that should be used to aggregate the
data. Default to 'ordem'.}

\item{spp.name}{the name of the columns containing the (morpho)species names.
Default to the TreeCo column 'species.correct'.}

\item{trait.list}{a vector with the names of the traits should be included in
the calculations.}

\item{rm.flora}{should the records of (morpho)species not found within the
main sampling but found in the floristics or natural regeneration of the
same survey. Default to TRUE.}

\item{trait.miss}{cutoff of number of missing traits}
}
\value{
a data frame with number of individuals (N) and taxa (S) for the each
  group and the percentage of N and S with all selected traits (columns
  'N.traits' and 'S.traits'), as well as a metric of the completeness of the
  traits within each group (columns 'miss.traits.N' and 'miss.traits.S').
}
\description{
Calculate the number of individuals and taxa per survey that
have any information available for selected species traits.
}
\details{
For each selected group, the function calculates the percentage of
  individuals and species that have information for all the traits listed in
  `trait.list`. This can be used as a measure of the trait coverage within
  each group.
  
  The function also returns a metric of how complete the trait information is
  for each group (i.e. survey). This metric is obtained by assigned for each
  taxa the number of missing traits, which goes from zero (all traits are
  available) to the maximum number of traits listed in `trait.list`. These
  number of missing traits are standardized by the maximum number of traits,
  so they vary between 0 and 1. Finally, these values are weighted by the number
  of individuals/taxa of each level of missing traits. That is, this metric
  is the weighted average of the number of missing traits per taxa. The closer to zero,
  more complete the taxa are in terms of the traits available.
  
  Some TreeCo surveys contain extra species lists from outside the sampling
  plots/transects. These lists are often obtaining from floristic surveys in
  the same sites where the sampling was conducted. Other surveys stored in
  TreeCo present the records for species present in the adult and natural
  regeneration components of the forest. If a species is present in the
  floristic survey but not in the plots, the species enters the database with
  an abundance value equal to zero. Similarly, if a species is present in the
  natural regeneration but not in the adult component, the species enters the
  database with an abundance between zero and one. This records are excluded by
  default, using the argument `rm.flora`.
}
\seealso{
\link[TreeCo]{indetSpecies}
}
\author{
Renato A. F. de Lima
}
