% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tspecies}
\alias{tspecies}
\title{tspecies dataset}
\format{A data frame with 3178 rows and 11 variables:
\describe{
  \item{code}{The species code}
  \item{botanical}{Botanical tree name}
  \item{common}{Common tree name}
  \item{region}{Tree region}
  \item{type}{Tree type}
  \item{assignment}{Another tree code for linking trees to benefits}
  \item{rating}{Percent species rating}
  \item{price_sq_in}{Price per square inch}
  \item{palm_trunk_cost_ft}{Palm trunk cost per feet}
  \item{replacement_cost}{Tree replacement cost}
  \item{t_ar_sq_inches}{to be determined...}
  ...
}}
\source{
\url{https://www.itreetools.org}
}
\usage{
tspecies
}
\description{
A dataset containing species data and other attributes.
}
\keyword{datasets}
