% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.stats.R
\name{weighted.stats}
\alias{weighted.stats}
\title{Weighted Mean, Variance and Standard Deviation}
\usage{
weighted.stats(x, w, na.rm = FALSE, stats = "mean")
}
\arguments{
\item{x}{numerical. A vector of values to compute the weighted statistics.}

\item{w}{numerical. A vector of weights with the same length as \code{x}.}

\item{na.rm}{logical. Should \code{NA} values be removed? Default to FALSE.}

\item{stats}{character. Which weighted statistic should be returned: "mean",
"var" or "sd"?}
}
\description{
Calculates the weighted mean and the variance or standard
  deviation for a given weighted mean.
}
\details{
Weigths should vary between 0 and 1. If they don't, they are
  normalized to sum one. Weights equal to 0 means that the corresponding
  values in \code{x} will be omitted from the from the calculation.
}
\examples{

x <- c(1,3,4,8)
w <- c(0.2,0.5,0.7,0.9)
mean(x); weighted.stats(x, w)
sd(x); weighted.stats(x, w, stats = "sd")

}
\author{
Renato A. F. de Lima
}
\keyword{internal}
