% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benefits}
\alias{benefits}
\title{Benefits data}
\format{A data frame with 7440 rows and 12 variables:
\describe{
  \item{id}{id}
  \item{species_code}{species code}
  \item{x3_81}{dbh range}
  \item{x11_43}{dbh range}
  \item{x22_86}{dbh range}
  \item{x38_10}{dbh range}
  \item{x53_34}{dbh range}
  \item{x68_58}{dbh range}
  \item{x83_82}{dbh range}
  \item{x99_06}{dbh range}
  \item{x114_30}{dbh range}
  \item{species_region}{species region}
  \item{benefit}{benefit type}
  \item{unit}{unit of measurment for benefit}
  ...
}}
\source{
\url{https://www.itreetools.org}
}
\usage{
benefits
}
\description{
A dataset containing the benefits and other attributes.
}
\keyword{datasets}
