% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.cat2mean_EG}
\alias{.cat2mean_EG}
\title{Convert Ecological Groups to Ordinal Values}
\usage{
.cat2mean_EG(x)
}
\arguments{
\item{x}{a vector of ecologial groups}
}
\value{
the character `x` with the values for the categories
}
\description{
Convert ecological groups (e.g. pioneer, early secondary, etc.)
into an ordinal variable
}
\keyword{internal}
