% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.cat2mean_LT}
\alias{.cat2mean_LT}
\title{Convert Leaf Rigidity to Ordinal Values}
\usage{
.cat2mean_LT(x)
}
\arguments{
\item{x}{a vector of leaf rigidity categories}
}
\value{
the character `x` with the values for the categories
}
\description{
Convert leaf rigidity categories into an ordinal variable 
  
  -  membranaceous = "1"
  -  submembranaceous =  "1.35"
  -  subpapyraceous = "1.7"
  -  papyraceous =  "2" # lembra papiro
  -  pergamentaceous = "2.35"  # consistencia de pergaminho
  -  subchartaceous = "2.7"
  -  chartaceous = "3" #consistência de cartolina
  -  rigid chartaceous = "3.35"
  -  subcoriaceous = "3.7"
  -  "coriaceous" = "4" # lembra couro
  -  rigid coriaceous = "4.35"
}
\keyword{internal}
