% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.cat2mean_wsg}
\alias{.cat2mean_wsg}
\title{Convert Categorical Wood Density to Means}
\usage{
.cat2mean_wsg(x)
}
\arguments{
\item{x}{a vector of wood density categories, in portuguese}
}
\value{
the character `x` with the mean values for the categories
}
\description{
The Forest Products Laboratory (1955) classifies wood in
  respect to their basic density as: light (0,30-0,36g/cm³), moderately heavy
  (0,42-0,50 g/cm³), heavy (0,50-0,60g/cm³) and vey heavy (0,60-0,72 g/cm³).
  
  Wong (2002) and Bruzos (2009) classify wood density as light (<0.5 g/cm3),
  moderately heavy (0.5-0.8 g/cm3), heavy (between 0.8-1.0 g/cm3) and very
  heavy (>1.0 g/cm3).
  
  In TreeCo, these categories of wood density mostly comes from H. Lorenzi's
  books, which were adpated here as follows:
  - muito leve: 0.30 g/cm³
  - leve: 0.45 g/cm³
  - moderadamente pesada: 0.65 g/cm³
  - pesada: 0.80 g/cm³
  - muito pesada: 0.95 g/cm³
}
\references{
Bruzos, T. (2009) Propiedades físicas de la madera. Maderas: Ciencia y
tecnología. Universidad del Bio, Chile. 11(1):3-18.

Wong, T. M. (2002). A Dictionary of Malaysian Timbers. Revised by Lim, S. C.
& Chung, R. C. K. Malayan Forest Record; No. 30. Forest Research Institute
Malaysia. Printed in Malaysia by Percetakan Haji Jantan, Kuala Lumpur,
Malaysia. pp.201.
}
\keyword{internal}
