% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco_data}
\alias{eco_data}
\title{Eco data}
\format{A data frame with 34164 rows and 5 variables:
\describe{
  \item{id}{id number}
  \item{species_code}{species code of tree}
  \item{species_region}{region of tree}
  \item{dbh_range}{dbh ranges for tree to fall under}
  \item{benefit_value}{benefit value associate with each dbh range}
  \item{benefit}{benefit type}
  ...
}}
\source{
\url{https://www.itreetools.org}
}
\usage{
eco_data
}
\description{
A dataset containing the benefits and other attributes.
}
\keyword{datasets}
