% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_guess.R
\name{eco_guess}
\alias{eco_guess}
\title{Guess common or botanical names}
\usage{
eco_guess(x, guess)
}
\arguments{
\item{x}{A vector containing "common" or "botanical" names.}

\item{guess}{The missing field, either "common" or "botanical".}
}
\description{
This function tries to guess the botanical or common name of
tree for users with only one of these fields, not both. Since \code{eco_run}
attempts to find matches in both common and botanical, running this can
increase the number of records you can grab to run eco benefits.
}
