% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_run.R
\name{eco_run}
\alias{eco_run}
\title{Run eco benefits for a tree}
\usage{
eco_run(species, dbh, region)
}
\arguments{
\item{species}{species common name, see \code{species_data}}

\item{dbh}{dbh value, any positive number}

\item{region}{region code, see \code{species_data} or \code{eco_data}}
}
\description{
Run eco benefits for a tree
}
