% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_run.R
\name{eco_run}
\alias{eco_run}
\title{Eco benefits calculator}
\usage{
eco_run(common, dbh, region)
}
\arguments{
\item{common}{the common name of the tree}

\item{dbh}{the dbh value in inches}

\item{region}{region code, see \code{money} dataset for region codes}
}
\description{
Run the eco benefits for a single tree. This function acts like
a benefits calculator in that you give a common name and it calculates the
benefits. Similar to Davey Tree's benefits calculator.
}
\examples{
eco_run("common fig", 20, "InlEmpCLM")
eco_run("red maple", 15, "PacfNWLOG")

}
