% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_run_all.R
\name{eco_run_all}
\alias{eco_run_all}
\title{Eco benefits for multiple trees}
\usage{
eco_run_all(data, common_col, botanical_col, dbh_col, region, n = 0.8,
  print_time = NULL)
}
\arguments{
\item{data}{path to csv file containing tree inventory}

\item{common_col}{the name of the column containing common names}

\item{botanical_col}{the name of the column containing botanical names}

\item{dbh_col}{the name of the column containing dbh values}

\item{region}{region code, see \code{species} or \code{benefits}}

\item{n}{guessing threshold from 0.0 to 1.0, defaults at 0.8}

\item{print_time}{Logical TRUE or FALSE for printing the elapsed time}
}
\description{
This function calculates the benefits for an entire dataset,
sometimes called a tree inventory. This works similar to i-Trees Eco software
where the user supplies data, the common name field, the botanical name
field, the dbh field, and the region. All calculations are done on unique
species/dbh pairs to avoid redundant computation and speed up the
calculation.
}
