% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_run_all.R
\name{eco_run_all}
\alias{eco_run_all}
\title{Run eco benefits for an entire tree inventory}
\usage{
eco_run_all(data, common_col, botanical_col, dbh_col, region,
  print_time = NULL)
}
\arguments{
\item{data}{path to csv file containing tree inventory}

\item{common_col}{the name of the column containing common names}

\item{botanical_col}{the name of the column containing botanical names}

\item{dbh_col}{the name of the column containing dbh values}

\item{region}{region code, see \code{species} or \code{benefits}}

\item{print_time}{Logical TRUE or FALSE for printing the elapsed time}
}
\description{
Run eco benefits for an entire tree inventory
}
